% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRisk_HBfn.R
\name{deepRiskHB}
\alias{deepRiskHB}
\title{Estimation of Risk Preference Parameters Using Hierarchical Markov Chain Mote Carlo Bayesian Framework}
\usage{
deepRiskHB()
}
\value{
Outputs multiple files with Risk Preference Parameters.
}
\description{
\code{deepRiskHB} Estimates Risk Preferences and outputs results in current working directory.
}

