% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepStart.R
\name{deepRun}
\alias{deepRun}
\title{Runs the DEEP Transformation, DEEP Split and Hierarchical Bayes algorithm.}
\usage{
deepRun(DEEPtype, WD = getwd(), file_path = NULL)
}
\arguments{
\item{DEEPtype}{character string that specifies whether output is DEEP "risk" or "time".}

\item{WD}{allows the user to specify a working directory. Uses the current directory if no directory is specified.}

\item{file_path}{contains the file path to the Limesurvey/Qualtrics output.}
}
\description{
\code{deepRun} Takes in a filepath for the Qualtrics/Limesurvey output
along with a working directory and the DEEP type specification. If no working
directory is provided, it defaults to the current directory. If no file_path is provided
it will request that one ge given using the GUI.
}
\examples{
\dontrun{
deepRun(DEEPtype = "risk", file_path = "/MyDocuments/myQualtricsOutput.csv")
deepRun(DEEPtype = "risk", WD = "/MyCurrentProject/DEEP")
deepRun(DEEPtype = "risk", WD = "/MyCurrentProject/DEEP", file_path = "/MyDocuments/myQualtricsOutput.csv")
}

}

