% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_basefns.R
\name{deepSplit}
\alias{deepSplit}
\title{Splits Participant Data into Hierarchical Bayes inputs}
\usage{
deepSplit(raw_data, DEEPtype)
}
\arguments{
\item{raw_data}{dataframe containing DEEP risk/time output}

\item{DEEPtype}{character string that specifies whether output is DEEP "risk" or "time".}
}
\value{
Writes four CSV files named according to what DEEP type was specified.
}
\description{
\code{deepSplit} subsets the "transformed" Limesurvey or Qualtrics DEEP output, after it has been loaded
to a dataframe, and exports four CSV files into the current working directory that will be read by the Hierarchical Bayes.
Note: Correct specification of the DEEP type is required for proper naming of files.
}
\examples{
\dontrun{
deepSplit(rawdata = = df, DEEPtype = "risk")
}

}

