% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_basefns.R
\name{strEndsWith}
\alias{strEndsWith}
\title{Reads file extension and helps Identify whether file is XML or CSV.}
\usage{
strEndsWith(haystack, needle)
}
\arguments{
\item{haystack}{string input, used for the file_path}

\item{needle}{string input, used for the file extension.}
}
\value{
boolean TRUE or FALSE to indicate if the file extension (the "needle") was found in the filepath (the "haystack").
}
\description{
\code{strEndsWith} searches for a substring ("needle") within a larger string ("haystack").
It returns a boolean TRUE/FALSE if it finds the substring within the string.
Note: if the substring contains more characters than the string, it defaults to FALSE.
}
\examples{
\dontrun{
strEndsWith(haystack = "myQualtricsOutput.csv", needle = "csv")
}

}

