% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTime_HBfn.R
\name{deepTimeHB}
\alias{deepTimeHB}
\title{Estimation of Time Preference Parameters Using Hierarchical Markov Chain Mote Carlo Bayesian Framework}
\usage{
deepTimeHB(iterUntilSave = 10000, iterSaved = 40000)
}
\value{
Outputs multiple files with Time Preference Parameters.
}
\description{
\code{deepTimeHB} Estimates Risk Preferences and outputs results in current working directory.
}
